feat: keychain-page 경로 추가
