+++ src/App.tsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import './App.css';
-import Main from '@pages/main';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { GoogleOAuthProvider } from '@react-oauth/google';
+import { AuthProvider } from 'api/context/auth-provider';
+import Frame from '@components/frame';
+import DashboardFrame from '@components/dashboard/layout/dashboard-frame';
+import Main from '@pages/main';
 import SignIn from '@pages/sign-in';
+import AuthSignIn from '@pages/sign-in/sign-in-email';
 import SignUp from '@pages/sign-up';
 import ReadTerms from '@pages/sign-up/content/read-terms';
 import Password from '@pages/sign-up/content/password';
 import ResetPassword from '@pages/reset-password';
-import Dashboard from '@pages/dashboard';
-import Frame from '@components/frame';
-import AuthSignIn from '@pages/sign-in/sign-in-email';
-import { AuthProvider } from 'api/context/auth-provider';
 import TempUpdatePasswordConfirm from '@pages/temp-update-password/verify';
 import TempUpdatePasswordSetting from '@pages/temp-update-password/setting';
-import DashboardFrame from '@components/dashboard-frame';
-import TestPage from '@pages/test';
+import Dashboard from '@pages/dashboard';
+import KeychainPage from '@pages/keychain';
 
 const clientId = process.env.REACT_APP_GOOGLE_CLIENT_ID || '';
 
@@ -35,10 +35,10 @@ const App: React.FC = () => {
               <Route path="read-terms" element={<ReadTerms />} />
               <Route path="temp-update-password-confirm" element={<TempUpdatePasswordConfirm />} />
               <Route path="temp-update-password-setting" element={<TempUpdatePasswordSetting />} />
-              <Route path="test-page" element={<TestPage />} />
             </Route>
             <Route path="/dashboard" element={<DashboardFrame />}>
               <Route index element={<Dashboard />} />
+              <Route path="keychain-page" element={<KeychainPage />} />
             </Route>
           </Routes>
         </BrowserRouter>
