+++ src/pages/keychain/index.style.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const KeychainContainer = styled.div`
+  display: flex;
+  gap: 40px;
+  flex-wrap: wrap;
+`;
